requirements.txt
# We will be using PyYAML module for parsing,  https://pyyaml.org/wiki/PyYAMLDocumentation
pyyaml
